feat(manage-ui): add session management ui
This PR implements a session management popup UI for existing sessions.
The popup is configurable and comes with "sensible" defaults (at least
to me).

The popup enables users to remove unnecessary or stale sessions in
nvim.oil fashion, just as simple as editing a buffer. After removing
stale sessions and quitting the popup, user will be prompted to confirm
changes.